Rename log to logger
